refactor in xray and dependency check driven adapters for get artifacts
