+++ tools/devsecops_engine_tools/engine_sca/engine_dependencies/test/infrastructure/driven_adapters/dependency_check/test_dependency_check_deserialize.py
@@ -14,7 +14,7 @@ def json_data():
     return {
         "dependencies": [
             {
-                "packages": [{"id": "package1"}],
+                "fileName": "path/to/package1:1.0",
                 "vulnerabilities": [
                     {
                         "name": "CVE-1234",
